requirements.txt
kafka-python==2.0.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6 